# Conflicts:
#	libraries/grpc-sdk/package.json
#	libraries/hermes/package.json
#	libraries/module-tools/src/metrics/MetricsServer.ts
#	modules/authorization/package.json
#	modules/chat/package.json
#	modules/database/package.json
#	modules/email/package.json
#	modules/forms/package.json
#	modules/functions/package.json
#	modules/push-notifications/package.json
#	modules/sms/package.json
#	modules/storage/package.json
#	packages/admin/package.json
#	packages/admin/src/middleware/Auth.middleware.ts
#	packages/commons/package.json
#	packages/core/package.json
#	yarn.lock

+++ modules/authentication/src/handlers/common.ts
@@ -3,18 +3,20 @@ import moment from 'moment';
 import ConduitGrpcSdk, {
   ConduitRouteActions,
   ConduitRouteReturnDefinition,
-  ConduitString,
-  ConfigController,
   GrpcError,
   ParsedRouterRequest,
-  RoutingManager,
   UnparsedRouterResponse,
 } from '@conduitplatform/grpc-sdk';
 import { status } from '@grpc/grpc-js';
 import { RefreshToken, User } from '../models';
-import { IAuthenticationStrategy } from '../interfaces/AuthenticationStrategy';
+import { IAuthenticationStrategy } from '../interfaces';
 import { Config } from '../config';
 import { TokenProvider } from './tokenProvider';
+import {
+  ConduitString,
+  ConfigController,
+  RoutingManager,
+} from '@conduitplatform/module-tools';
 
 export class CommonHandlers implements IAuthenticationStrategy {
   constructor(private readonly grpcSdk: ConduitGrpcSdk) {}
@@ -40,10 +42,18 @@ export class CommonHandlers implements IAuthenticationStrategy {
     if (!oldRefreshToken.user) {
       throw new GrpcError(status.PERMISSION_DENIED, 'Invalid user');
     }
+    if (!(oldRefreshToken.user as User).active) {
+      throw new GrpcError(status.PERMISSION_DENIED, 'User is blocked');
+    }
 
     // delete the old refresh token
     await RefreshToken.getInstance().deleteOne({ _id: oldRefreshToken._id });
-
+    // delete all expired tokens
+    RefreshToken.getInstance()
+      .deleteMany({
+        expiresOn: { $lte: new Date() },
+      })
+      .catch();
     return TokenProvider.getInstance().provideUserTokens({
       user: oldRefreshToken.user as User,
       clientId,
@@ -64,7 +74,7 @@ export class CommonHandlers implements IAuthenticationStrategy {
     await TokenProvider.getInstance().logOutClientOperations(
       this.grpcSdk,
       clientConfig,
-      authToken,
+      authToken!,
       clientId,
       user._id,
     );
